requirements.txt
google-cloud-speech
matplotlib
pyaudio